+++ lib/features/dashboard/dashboard_screen.dart
@@ -683,7 +683,7 @@ class DashboardScreen extends StatelessWidget {
                       const SizedBox(height: 32),
                       const Center(
                         child: Text(
-                          'HelpDesk Connect Beta v0.6.4 Build 16',
+                          'HelpDesk Connect Beta v0.6.4 Build 17',
                           style: TextStyle(color: HDCColors.textSecondary),
                         ),
                       ),